Continued Refactoring of JSON values (see Details)
Updated the refactoring of JSON into different categories.
Each set of options goes in this order: Display Options, Strings (if applicable), Color, Size, Other Options (if applicable)
Added steps and minimum values for all numerical values
Moved evidence up in list, so that the options go from top-to-bottom, left-to-right
Updated Conclusion Messages Group + Labels. The JS handles all blanks with the defaults already, so this should not be an issue.
Updated Border Labels to try to make them easier to understand.